spit expand resource test into individual tests for each resource
